Moved sleep to its own file
