+++ packages/commands/src/query/cctx.ts
@@ -3,9 +3,8 @@ import { ethers } from "ethers";
 import EventEmitter from "eventemitter3";
 import { z } from "zod";
 
-import { CrossChainTx } from "../../../../types/trackCCTX.types";
-import { fetchFromApi } from "../../../../utils/api";
-import { sleep } from "../../../../utils/cctx";
+import type { CrossChainTx } from "../../../../types/trackCCTX.types";
+import { fetchFromApi, sleep } from "../../../../utils";
 
 /**
  * Event map:
