+++ utils/index.ts
@@ -11,6 +11,7 @@ export * from "./keyPaths";
 export * from "./parseAbiValues";
 export * from "./parseJson";
 export * from "./safeAwait";
+export * from "./sleep";
 export * from "./solanaAccounts";
 export * from "./toHexString";
 export * from "./trackCCTX";
